requirements.txt
Django==1.9.4
South==1.0.2
argparse==1.2.1
decorator==4.0.9
funcsigs==0.4
ipython==4.1.2
ipython-genutils==0.1.0
mock==1.3.0
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
wsgiref==0.1.2
